requirements.txt
pymodbus==3.6.9
influxdb_client==1.44.0
python-dotenv==1.0.1